organized the tests and move them in the content and ui directories
