+++ mobile/__tests__/content/app.test.js
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import App from '../App';
+import App from '../../App';
 import renderer from 'react-test-renderer';
 
 jest.useFakeTimers()
