+++ mobile/__tests__/content/view_books.test.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import renderer from 'react-test-renderer';
-import ViewBooks from '../components/screens/ViewBooks';
+import ViewBooks from '../../components/screens/ViewBooks';
 
 test('view books content matches the snapshot', () => {
     /* Content for ViewBooks.js matches the snapshot */
