+++ mobile/__tests__/content/view_books_by_author.test.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import renderer from 'react-test-renderer';
-import ViewBooksByAuthor from '../components/screens/ViewBooksByAuthor';
+import ViewBooksByAuthor from '../../components/screens/ViewBooksByAuthor';
 
 test('view books by author content matches the snapshot', () => {
     /* Content for ViewBooksByAuthor.js matches the snapshot */
