+++ mobile/__tests__/content/view_books_by_subject.test.js
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import renderer from 'react-test-renderer';
-import ViewBooksBySubject from '../components/screens/ViewBooksBySubject';
+import ViewBooksBySubject from '../../components/screens/ViewBooksBySubject';
 
 test('view books by subject content matches the snapshot', () => {
     /* Content for ViewBooksBySubject.js matches the snapshot */
